requirements.txt
asgiref==3.6.0
attrs==21.3.0
dj-database-url==0.5.0
Django==4.1.7
django-cors-headers==3.10.1
django-on-heroku==1.1.2
django-polymorphic==3.1.0
django-rest-passwordreset==1.2.1
django-rest-polymorphic==0.1.10
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-http-client==3.3.7
pytz==2021.3
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
supermemo2==2.0.0
whitenoise==6.0.0
